modified pandemy.cpp, State.hpp, main.cpp Pandemy.hpp

+++ mandatory_sir_model/State.hpp
@@ -1,8 +1,8 @@
 #ifndef STATE_HPP
 #define STATE_HPP
 struct State {
-  double S = 0;
-  double I = 0;
-  double R = 0;
+  int S = 0;
+  int I = 0;
+  int R = 0;
 };
 #endif
